Update to use new strategy pattern and authService interface
